sync script
python script for copying, overwriting and deleting files/directories across two directories. new files will be copied, if the same file exists but a newer version exists, it will add to an overwrite list, same for files that dont exist anymore, after the list is complete it shows the list and asks if you want to go ahead with the delete/overwrite operations. sync.py should be run within the directory you are syncing to, and the source it copies from should be changed in the script.